refactor(billing): split upgrade-to-lifetime into its own router (#1011)
`backend/routes/subscriptions.py` was about to grow from 324 → 786 LOC
(+142%), tripping the godmodule LOC gate's growth warning. Move the
upgrade-to-lifetime endpoints, schemas and helper lookups into a new
`backend/routes/upgrade_to_lifetime.py` (522 LOC) and register it in
`startup/routes.py` next to `subscriptions_router`.

URL surface is unchanged: both routers share `prefix="/api/subscriptions"`
so `/v1/api/subscriptions/upgrade-to-lifetime[/preview]` still resolves
the same way.

Also sanitizes user-derived values flowing into log lines (CodeQL
py/log-injection — alerts 3513/3514/3515) by stripping CR/LF before
emitting.
